+++ src/Apecs/Core.hs
@@ -171,6 +171,8 @@ instance (ExplSet sa, ExplSet sb) => ExplSet (EitherStore sa sb) where
   explSet (EitherStore _ sb) ety (Right b) = explSet sb ety b
   explSet (EitherStore sa _) ety (Left a)  = explSet sa ety a
 
+instance Has w () where
+  getStore = return ()
 instance Component () where
   type Storage () = ()
 type instance Elem () = ()