+++ src/Apecs/Util.hs
@@ -11,7 +11,7 @@ module Apecs.Util (
   runGC, global,
 
   -- * EntityCounter
-  EntityCounter, nextEntity, newEntity,
+  EntityCounter(..), nextEntity, newEntity,
 
   -- * Spatial hashing
   -- $hash
@@ -26,7 +26,6 @@ import           Control.Applicative  (liftA2)
 import           Control.Monad.Reader (liftIO)
 import           Data.Monoid
 import           Data.Semigroup
-import           Data.Proxy
 import           System.CPUTime
 import           System.Mem           (performMajorGC)